+++ server/modules/admin/user/user.routes.js
@@ -1,0 +1,5 @@
+import AdminMerchantController from './user.controller';
+
+const { router } = AdminMerchantController.buildResource('merchants', ['index', 'new', 'create', 'edit', 'update'])
+
+export default router