+++ server/modules/lib/middleware/administrator-check.js
@@ -1,7 +1,0 @@
-
-export default async (req, _, next) => {
-    if (!req.user || await !req.user.isAdmin()) {
-        return next(new Error('Unauthorized Access'));
-    }
-    next();
-}