+++ server/modules/merchant/message/message.controller.js
@@ -1,17 +1,17 @@
-import MerchantController from '../merchant-controller';
+import AdminViewController from '../../lib/admin-view-controller';
 import { Message } from '../../../models';
 
-export default class MessageController extends MerchantController {
+export default class MessageController extends AdminViewController {
     getResourceName() {
         return 'message'
     }
 
-    async indexPage(req) {
+    async indexPage(req, res) {
         const messages = await Message.findAll({
             where: { catalogId: req.catalog.id },
             order: [['created_at', 'DESC']]
         });
         const page = await this._getMerchantPage(req.catalog.id, 'messages');
-        return { pageParams: { messages }, currentPage: page };
+        return super.indexPage({ res, pageParams: { messages }, currentPage: page });
     }
 }