+++ server/modules/merchant/review/review.routes.js
@@ -1,26 +1,5 @@
-import RouteBuilder from '../route-builder';
 import ReviewController from './review.controller';
 
-const router = RouteBuilder(ReviewController, (router) => {
-    router
-        .route('/reviews')
-        .get(router.indexPage);
-
-    router
-        .route('/reviews/new')
-        .get(router.newPage);
-
-    router
-        .route('/reviews')
-        .post(router.createAction);
-
-    router
-        .route('/reviews/:id/edit')
-        .get(router.editPage);
-
-    router
-        .route('/reviews/:id/update')
-        .put(router.updateAction);
-});
+const { router } = ReviewController.buildResource('reviews', ['index', 'new', 'create', 'edit', 'update']);
 
 export default router;